requirements.txt
slack_bolt
slack_sdk
schedule
psycopg2-binary
gunicorn
python-dotenv

